requirements.txt
# Dependências Python - API Aponta Supabase
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.24.1
commitizen==3.15.0

# Supabase
supabase==2.3.4

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Dependências de Teste
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-json-report==1.5.0
